Added checking for output file changes
This is required to handle cases where outputs are modified by some
other process, either by accident or intent.

This might potentially double the work done checking SHA256 of a
particular file if the output of one workflow is the input of another
(as is expected to be a common case). Perhaps there is some
cross-workflow-step optimization to be done in the future.